fix type

+++ crawl-ref/source/god-passive.cc
@@ -936,7 +936,7 @@ unsigned int ash_skill_point_boost(skill_type sk, int scaled_skill)
     unsigned int skill_points = 0;
 
     int skill = max(scaled_skill, 1);
-    skill_points += (you.skill_boost[sk] * 2 + 1) * isqrt(skill * skill * min(161, you.piety));
+    skill_points += (you.skill_boost[sk] * 2 + 1) * isqrt(skill * skill * min(161, (int)you.piety));
 
     return skill_points;
 }
